% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arf.R
\name{arf}
\alias{arf}
\title{Beta(a,b) distribution by the acceptance-rejection method}
\usage{
arf(n, a, b)
}
\arguments{
\item{n}{number}

\item{a}{number}

\item{b}{number}
}
\value{
times  \code{n}
}
\description{
the times of getting Beta(a,b) distribution by the acceptance-rejection method
}
\examples{
\dontrun{
n=1000
a=3
b=2
arf(1000,3,2)
}
}
