% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test2.R
\name{chisq.test2}
\alias{chisq.test2}
\title{Pearson's Chi-squared Test for Count Data}
\usage{
chisq.test2(x1, x2)
}
\arguments{
\item{x1}{the first vector,any length}

\item{x2}{the second vector,needing same length}
}
\value{
X-squared  \code{n}
}
\description{
chisq.test performs chi-squared contingency table tests and goodness-of-fit tests,the faster version
}
\examples{
\dontrun{
a=c(12,24)
b=c(25,10)
chisq2=chisq.test2(a,b)
chisq2
}
}
