% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arf.R
\name{arf}
\alias{arf}
\title{Beta(a,b) distribution by the acceptance-rejection method}
\usage{
arf(n, a, b)
}
\arguments{
\item{n}{the whole number of Beta distribution}

\item{a}{the first parameter of Beta distribution}

\item{b}{the second parameter of Beta distribution}
}
\value{
times  \code{n}
}
\description{
the required times of getting n points which satidfy Beta(a,b) distribution by the acceptance-rejection method
}
\examples{
\dontrun{
n=1000
a=3
b=2
arf(1000,3,2)
}
}
