% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test2.R
\name{chisq.test2}
\alias{chisq.test2}
\title{A chi-square test using R}
\usage{
chisq.test2(x1, x2)
}
\arguments{
\item{x1}{vector}

\item{x2}{vector}
}
\value{
X-squared  \code{n}
}
\description{
A chi-square test using R
}
\examples{
\dontrun{
a=c(12,24)
b=c(25,10)
chisq2=chisq.test2(a,b)
chisq2
}
}
