% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Cross Tabulation and Table Creation}
\usage{
table2(x, y)
}
\arguments{
\item{x}{the first vector,any length}

\item{y}{the second vector,needing same length}
}
\value{
a contingency table \code{n}
}
\description{
table2 uses the cross-classifying factors to build a contingency table of the counts at each combination of factor levels.a faster version.
}
\examples{
\dontrun{
a=c(0,1,2,1,1,0,0,1)
b=c(0,0,1,1,1,0,1,1)
table2(a,b)
}
}
